+++ tests/test_text/test_text_counting.py
